requirements.txt
wavelink
discord.py
